+++ tests/functional/ext/docparams/raise/missing_raises_doc.py
@@ -1,6 +1,6 @@
 """Tests for missing-raises-doc and missing-raises-type-doc"""
 # pylint: disable=function-redefined, invalid-name, undefined-variable, missing-function-docstring
-# pylint: disable=unused-argument, import-error
+# pylint: disable=unused-argument, import-error, unused-variable
 
 from unknown import Unknown
 
@@ -20,4 +20,21 @@ def test_ignores_raise_uninferable(self):
     :raises NameError: Never
     """
     raise Unknown("hi")
+    raise NameError("hi")  # [unreachable]
+
+
+def test_ignores_returns_from_inner_functions(self):  # [missing-raises-doc]
+    """This is a docstring.
+    We do NOT expect a warning about the OSError in inner_func!
+
+    :raises NameError: Never
+    """
+
+    def ex_func(val):
+        def inner_func(value):
+            return OSError(value)
+
+        return RuntimeError(val)
+
+    raise ex_func("hi")
     raise NameError("hi")  # [unreachable]
